# Add CLI startup profiling feature