--- a/packages/cli/src/commands/test-profile.ts
+++ b/packages/cli/src/commands/test-profile.ts
@@ -1,4 +1,5 @@
 import { CommandModule } from 'yargs';
+
 import { SharedOptions } from '../options.js';
 
 export const command: CommandModule<object, SharedOptions> = {
@@ -7,8 +8,8 @@ export const command: CommandModule<object, SharedOptions> = {
   handler: async () => {
     console.log('Profile test completed successfully');
     // Profiling report will be automatically displayed by the main function
-    
+
     // Force a delay to simulate some processing
-    await new Promise(resolve => setTimeout(resolve, 100));
+    await new Promise((resolve) => setTimeout(resolve, 100));
   },
 };