--- a/packages/cli/src/utils/performance.ts
+++ b/packages/cli/src/utils/performance.ts
@@ -25,22 +25,26 @@ export function mark(label: string): void {
  */
 export async function reportTimings(): Promise<void> {
   if (!isEnabled) return;
-  
-  console.log('\n📊 Performance Profile:');
+
+  console.log('\n📊 Performance Profile Results');
   console.log('=======================');
-  
+  console.log(
+    `${'Label'.padEnd(40, ' ')}${'Time'.padStart(10, ' ')}${'Duration'.padStart(10, ' ')}`,
+  );
+
   // Sort timings by time value
-  const sortedTimings = Object.entries(timings)
-    .sort((a, b) => a[1] - b[1]);
-  
+  const sortedTimings = Object.entries(timings).sort((a, b) => a[1] - b[1]);
+
   // Calculate durations between steps
   let previousTime = 0;
   for (const [label, time] of sortedTimings) {
     const duration = time - previousTime;
-    console.log(`${label}: ${time.toFixed(2)}ms (${duration.toFixed(2)}ms)`);
+    console.log(
+      `${label.padEnd(40, ' ')}${`${time.toFixed(2)}ms`.padStart(10, ' ')}${`${duration.toFixed(2)}ms`.padStart(10, ' ')}`,
+    );
     previousTime = time;
   }
-  
+
   console.log(`Total startup time: ${previousTime.toFixed(2)}ms`);
   console.log('=======================\n');
 
@@ -55,32 +59,35 @@ export async function reportTimings(): Promise<void> {
  */
 async function reportPlatformInfo(): Promise<void> {
   if (!isEnabled) return;
-  
+
   console.log('\n🖥️ Platform Information:');
   console.log('=======================');
   console.log(`Platform: ${process.platform}`);
   console.log(`Architecture: ${process.arch}`);
   console.log(`Node.js version: ${process.version}`);
-  
+
   // Windows-specific information
   if (process.platform === 'win32') {
     console.log('Windows-specific details:');
     console.log(`- Current working directory: ${process.cwd()}`);
     console.log(`- Path length: ${process.cwd().length} characters`);
-    
+
     // Check for antivirus markers by measuring file read time
     try {
       // Using dynamic import to avoid require
       const fs = await import('fs');
       const startTime = performance.now();
       fs.readFileSync(process.execPath);
-      console.log(`- Time to read Node.js executable: ${(performance.now() - startTime).toFixed(2)}ms`);
+      console.log(
+        `- Time to read Node.js executable: ${(performance.now() - startTime).toFixed(2)}ms`,
+      );
     } catch (error: unknown) {
-      const errorMessage = error instanceof Error ? error.message : String(error);
+      const errorMessage =
+        error instanceof Error ? error.message : String(error);
       console.log(`- Error reading Node.js executable: ${errorMessage}`);
     }
   }
-  
+
   console.log('=======================\n');
 }