--- a/packages/cli/tests/settings/config.test.ts
+++ b/packages/cli/tests/settings/config.test.ts
@@ -44,6 +44,7 @@ describe('Config', () => {
         modelProvider: 'anthropic',
         modelName: 'claude-3-7-sonnet-20250219',
         ollamaBaseUrl: 'http://localhost:11434/api',
+        profile: false,
         customPrompt: '',
       });
       expect(fs.existsSync).toHaveBeenCalledWith(mockConfigFile);
@@ -77,6 +78,7 @@ describe('Config', () => {
         modelProvider: 'anthropic',
         modelName: 'claude-3-7-sonnet-20250219',
         ollamaBaseUrl: 'http://localhost:11434/api',
+        profile: false,
         customPrompt: '',
       });
     });