@bhouston 2025-03-05
I've added the `profile` option to the persistent config file as requested. This allows you to enable profiling permanently on your development machine.

### How to Use

You can enable profiling in your config with:

```bash
# Enable profiling in config
mycoder config set profile true

# Disable profiling in config
mycoder config set profile false
```

Once enabled in the config, all `mycoder` commands will automatically show profiling information without needing the `--profile` flag.